08/07 Исправлено: - Считывание большого количества данных. - считывание с бд массива данных температуры. - отображение графика температуры; - Добавлено клонирование класса StoreDataModel

+++ app/src/main/java/com/project/tlogger/msg/Lib.java
@@ -19,7 +19,7 @@ public class Lib {
     public int runningTime;
 
     public short count;
-    public byte measurementsCount;
+    public short measurementsCount;
 
     public boolean setConfigurationFlag = false;
